githubctl/pyproject.toml
[tool.poetry]
name = "githubctl"
version = "0.1.0"
description = ""
authors = ["Felipe Padilha <62897335+padilhafe@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = {extras = ["all"], version = "^0.13.1"}
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
